C-Sharp Code
Coding challenge
Create a simple Contacts Web app – backend and frontend in C#
1. Use .Net/.Net Core framework
2. Use any your favorite db (ms sql server, mysql, postgres, in-memory
db etc)
3. Create table tblPersons (id, first_name, last_name, phone)
4. Init with some initial data
5. Use your favorite ORM (Entity Framework, Drapper etc…)
6. Create WebAPI interface for tblPersons (get, put, delete, post) or in
traditional MVC model
7. Create automatic test functions (XUnit, NUnit, MS Tests framework
etc) for some functions
8. Create frontend in your favorite framework (js, vue, react, angular -
in case of WebAPI,Razor in case of MVC) in order to
a.  show list of persons and allow to edit, create and delete record
UI/UX is not important in that coding challenge so make it simple.